Refactor ui module
