+++ src/actions/delete.rs
@@ -1,5 +1,8 @@
 use crate::actions::{handle_matches, ItemType, MatchHandlerTemplate, UnlockingAction};
 use crate::ui;
+use crate::ui::output::{
+    show_credentials_table, show_notes_table, show_payment_cards_table, show_totp_table,
+};
 use crate::vault::entities::{Credential, Error, Note, PaymentCard, Totp};
 use crate::vault::vault_trait::Vault;
 use clap::ArgMatches;
@@ -25,7 +28,7 @@ impl<'a> MatchHandlerTemplate for DeleteCredentialsTemplate<'a> {
         &mut self,
         matches: Vec<Self::ItemType>,
     ) -> Result<Option<String>, Error> {
-        ui::show_credentials_table(&matches, false);
+        show_credentials_table(&matches, false);
         match ui::ask_index(
             "To delete, please enter a row number from the table above, press a to delete all, or press q to abort",
             matches.len() as i16 - 1,
@@ -56,7 +59,7 @@ impl<'a> MatchHandlerTemplate for DeletePaymentTemplate<'a> {
 
     fn pre_handle_matches(&self, matches: &Vec<Self::ItemType>) {
         println!("Found {} payment cards...", matches.len());
-        ui::show_payment_cards_table(matches, false);
+        show_payment_cards_table(matches, false);
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
@@ -100,7 +103,7 @@ impl<'a> MatchHandlerTemplate for DeleteNoteTemplate<'a> {
 
     fn pre_handle_matches(&self, matches: &Vec<Self::ItemType>) {
         println!("Found {} notes", matches.len());
-        ui::show_notes_table(matches, false);
+        show_notes_table(matches, false);
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
@@ -145,7 +148,7 @@ impl<'a> MatchHandlerTemplate for DeleteTotpTemplate<'a> {
 
     fn pre_handle_matches(&self, matches: &Vec<Self::ItemType>) {
         println!("Found {} TOTP entries", matches.len());
-        ui::show_totp_table(matches);
+        show_totp_table(matches);
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
