+++ src/actions/mod.rs
@@ -5,6 +5,7 @@ pub mod export;
 pub mod generate;
 pub mod help;
 pub mod import;
+pub mod init;
 pub mod lock;
 pub mod show;
 pub mod unlock;
