+++ src/actions/show.rs
@@ -2,6 +2,10 @@ use crate::actions::{
     copy_to_clipboard, handle_matches, ItemType, MatchHandlerTemplate, UnlockingAction,
 };
 use crate::ui;
+use crate::ui::output::{
+    show_card, show_credentials_table, show_note, show_notes_table, show_payment_cards_table,
+    show_totp_table,
+};
 use crate::vault::entities::{Credential, Error, Note, PaymentCard, Totp};
 use crate::vault::vault_trait::Vault;
 use clap::ArgMatches;
@@ -23,7 +27,7 @@ impl MatchHandlerTemplate for ShowCredentialsTemplate {
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
-        ui::show_credentials_table(&vec![the_match.clone()], self.verbose);
+        show_credentials_table(&vec![the_match.clone()], self.verbose);
         copy_to_clipboard(the_match.password());
         Ok(Some("Password copied to clipboard!".to_string()))
     }
@@ -32,7 +36,7 @@ impl MatchHandlerTemplate for ShowCredentialsTemplate {
         &mut self,
         matches: Vec<Self::ItemType>,
     ) -> Result<Option<String>, Error> {
-        ui::show_credentials_table(&matches, self.verbose);
+        show_credentials_table(&matches, self.verbose);
 
         match ui::ask_index(
             "To copy one of these passwords to clipboard, please enter a row number from the table above, or press q to exit",
@@ -61,11 +65,11 @@ impl MatchHandlerTemplate for ShowPaymentsTemplate {
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
-        ui::show_payment_cards_table(&vec![the_match.clone()], self.show_cleartext);
+        show_payment_cards_table(&vec![the_match.clone()], self.show_cleartext);
         copy_to_clipboard(the_match.number());
         match ui::ask("Do you want to see the full card details? (y/n)").as_str() {
             "y" => {
-                ui::show_card(&the_match);
+                show_card(&the_match);
                 Ok(Some("Card number copied to clipboard!".to_string()))
             }
             _ => Ok(Some("Card number copied to clipboard!".to_string())),
@@ -76,14 +80,14 @@ impl MatchHandlerTemplate for ShowPaymentsTemplate {
         &mut self,
         matches: Vec<Self::ItemType>,
     ) -> Result<Option<String>, Error> {
-        ui::show_payment_cards_table(&matches, self.show_cleartext);
+        show_payment_cards_table(&matches, self.show_cleartext);
 
         match ui::ask_index(
             "To see card details, enter a row number from the table above, or press q to exit",
             matches.len() as i16 - 1,
         ) {
             Ok(index) => {
-                ui::show_card(&matches[index]);
+                show_card(&matches[index]);
                 copy_to_clipboard(matches[index].number());
                 Ok(Some("Card number copied to clipboard!".to_string()))
             }
@@ -104,10 +108,10 @@ impl MatchHandlerTemplate for ShowNotesTemplate {
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
-        ui::show_notes_table(&vec![the_match.clone()], self.verbose);
+        show_notes_table(&vec![the_match.clone()], self.verbose);
         let response = ui::ask("Do you want to see the full note? (y/n)");
         if response == "y" {
-            ui::show_note(&the_match);
+            show_note(&the_match);
         }
         Ok(None)
     }
@@ -116,14 +120,14 @@ impl MatchHandlerTemplate for ShowNotesTemplate {
         &mut self,
         matches: Vec<Self::ItemType>,
     ) -> Result<Option<String>, Error> {
-        ui::show_notes_table(&matches, self.verbose);
+        show_notes_table(&matches, self.verbose);
 
         match ui::ask_index(
             "To see the full note, please enter a row number from the table above, or press q to exit",
             matches.len() as i16 - 1,
         ) {
             Ok(index) => {
-                ui::show_note(&matches[index]);
+                show_note(&matches[index]);
                 Ok(None)
             }
             Err(message) => {
@@ -140,7 +144,7 @@ impl MatchHandlerTemplate for ShowTotpTemplate {
 
     fn pre_handle_matches(&self, matches: &Vec<Self::ItemType>) {
         println!("Found {} matching OTP authorizers:", matches.len());
-        ui::show_totp_table(matches);
+        show_totp_table(matches);
     }
 
     fn handle_one_match(&mut self, the_match: Self::ItemType) -> Result<Option<String>, Error> {
